Add files via upload
Added 2 new Arduino Uno sketches, both testing of Maxbotix EZ MB1040 and  the HRLV MB1013 MB1023 MB1033 sensors. One sketch does Pulse duration measurements, and one sketch does Analog Voltage measurements. 
Note both families require different scaling factors, so there is a #if def type switch in the loop or header section, to help you define if you use a EZ or a HRLV family sensor.
In my case the Voltage Measurement method required customization of the scaling factor possibly due to the voltage measurement of my particular UNO ??? or perhaps documentation was incorrect. But the pulse duration method was spot on, no calibration required, just you must still change the #if 1 to a #if 0 if you are using the HRLV sensor. Read about #ifdef type compiler messages if you are not sure.